pyproject.toml
[tool.poetry]
name = "horizons_client"
version = "0.1.9"
description = "An API Client for communicating with JPL's Horizons Ephemeride database."
authors = ["alevan <levan.andrew.w@gmail.com>"]
license = "MIT"

[[tool.poetry.source]]
name = "home"
url = "http://pypi.home/"

[tool.poetry.dependencies]
python = "^3.8"
pandas = "1.3.5"
python-dateutil = "2.8.2"
aiohttp = ">=3.7.4.post0"
twine = "^3.8.0"
asynctest = "^0.13.0"

[tool.poetry.dev-dependencies]
pytest = "6.2.5"
pytest-freezegun = "0.4.2"
pytest-lazy-fixture = "0.6.3"
pytest-asyncio = "0.16.0"
pylint = "2.12.2"
black = "22.3.0"
pre-commit = "2.16.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
